+++ src/telegram/actions/common.ts
@@ -6,8 +6,9 @@ import {
   type TelegramApi,
 } from "../api/tgapi.js";
 import type { TgMessage } from "../api/types.js";
-import type { MessageOptions, TelegramMessagePrefix } from "../types.js";
+import type { MessageOptions } from "../types.js";
 import type { BotMessageModel } from "../models/botMessage.js";
+import type { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import type { getDb } from "../../db/index.js";
 import type { LanguageCode } from "../../recognition/types.js";
 import type { TranslationKeyFull } from "../../text/types.js";
