+++ src/telegram/actions/donate.ts
@@ -1,8 +1,9 @@
 import { GenericAction } from "./common.js";
-import { TelegramButtonModel, TelegramMessagePrefix } from "../types.js";
+import { TelegramButtonModel } from "../types.js";
 import { BotCommand } from "../commands.js";
 import { isCommandMessage } from "../commandsChecker.js";
 import { BotMessageModel } from "../models/botMessage.js";
+import { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import { getBotLogo, getPaymentInvoiceData, isStars } from "../helpers.js";
 import { TranslationKeys } from "../../text/types.js";
 import { Logger } from "../../logger/index.js";
