+++ src/telegram/actions/ignore.ts
@@ -1,7 +1,7 @@
 import { GenericAction } from "./common.js";
 import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
-import type { TelegramMessagePrefix } from "../types.js";
+import type { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import type { BotMessageModel } from "../models/botMessage.js";
 import type { TgMessage } from "../api/types.js";
