+++ src/telegram/actions/lang.ts
@@ -1,11 +1,8 @@
 import { GenericAction } from "./common.js";
-import {
-  BotLangData,
-  TelegramButtonModel,
-  TelegramMessagePrefix,
-} from "../types.js";
+import { BotLangData, TelegramButtonModel } from "../types.js";
 import { BotCommand } from "../commands.js";
 import { isCommandMessage } from "../commandsChecker.js";
+import { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import { BotMessageModel } from "../models/botMessage.js";
 import { getLanguageByText, getRawUserLanguage } from "../helpers.js";
 import { type TranslationKey, TranslationKeys } from "../../text/types.js";
