+++ src/telegram/actions/start.ts
@@ -4,7 +4,7 @@ import { isCommandMessage } from "../commandsChecker.js";
 import { TranslationKeys } from "../../text/types.js";
 import { Logger } from "../../logger/index.js";
 import { collectAnalytics } from "../../analytics/index.js";
-import { type TelegramMessagePrefix } from "../types.js";
+import type { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import type { BotMessageModel } from "../models/botMessage.js";
 import type { TgMessage } from "../api/types.js";
 import { trackUserActivity } from "../../monitoring/newrelic.js";
