+++ src/telegram/actions/voice.ts
@@ -9,7 +9,8 @@ import {
 import { collectAnalytics } from "../../analytics/index.js";
 import { TimeMeasure } from "../../common/timer.js";
 import { isBlockedByUser } from "../api/tgerror.js";
-import { type TelegramMessagePrefix, VoiceContentReason } from "../types.js";
+import { VoiceContentReason } from "../types.js";
+import type { TelegramMessagePrefix } from "../models/messagePrefix.js";
 import type { BotMessageModel } from "../models/botMessage.js";
 import { getFullFileName } from "../../files/index.js";
 import type { TgMessage } from "../api/types.js";
